r/virtual_machine_scale_set: switching to use a State Upgrader
